refactor to allow module execution
